// src/types/forms/promotion/externalPromotion/externalPromotionTypes.ts
export const defaultValues = {
    id: "",
    code: "Auto Generated",
    name: "",
    promotionType: [],
    promotionTypes: [],
    altName: "",
    externalConditionNumber: "",
    externalSalesConditionNumber: "",
    offerDescription: "",
    priority: "",
    active: true,
    internalCustomer: true,
    tenantId: "",
    companyId: "",
    // createdOn: "",
    createdBy: "",
    // updatedOn: "",
    updatedBy: "",
    externalReference: "",
  };

  export const defaultRecordsValues = {
    companyId: "",
    legalEntityId: "",
    promotionType: "",
    externalPromotionHeader: "",
  }
  